Rawsocket polimorphic implementation.
