+++ lib/avtp_pipeline/platform/Linux/x86_i210_linux.cmake
@@ -30,3 +30,4 @@ set ( PLATFORM_LINK_LIBRARIES
 #set ( PLATFORM_DEFINE "AVB_DELAY_TWEAK_USEC=15" )
 
 set ( GSTREAMER_1_0 1 )
+set ( AVB_FEATURE_PCAP 1 )
