+++ lib/avtp_pipeline/rawsock/CMakeLists.txt
@@ -1,1 +1,4 @@
-# No common code
+SET (SRC_FILES ${SRC_FILES}
+	${AVB_SRC_DIR}/rawsock/rawsock_impl.c
+	PARENT_SCOPE
+)
